FIX (findbugs): in meanQuality() fix divide cast.
git-svn-id: https://subversion.biologie.ens.fr/svn/leburon/trunk/eoulsan@1501 5a62d92b-e016-0410-9103-b89fe532bb34

+++ src/main/java/fr/ens/transcriptome/eoulsan/bio/ReadSequence.java
@@ -99,7 +99,7 @@ public class ReadSequence extends Sequence {
     for (int i = 0; i < len; i++)
       score += quality.charAt(i) - 64;
 
-    return score / len;
+    return score / (double) len;
   }
 
   /**
